requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
Babel==2.12.1
cffi==1.15.1
Django==4.1.9
django-cors-headers==3.14.0
django-sse==0.4.1
gevent==22.10.2
greenlet==2.0.2
markdown2==2.4.8
Pillow==9.4.0
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
toml==0.10.2
tzdata==2023.3
zope.event==4.6
zope.interface==6.0
